Adding inspire links to records
This should add a citations link with the inspire logo to the records.

Not sure if it's desirable to check if there are citations before
providing the link - because of this formatting, it's no overhead to
just have the link, but checking for citations would create overhead in
generating the pages.

Using the inspire favored icon for the image, happy to change if there's
a preferred icon or logo that anyone is aware of.

This is a quick attempt to address https://github.com/cernopendata/opendata.cern.ch/issues/3659