build(docker): upgrade cernopendata-portal to 0.13.0
Upgrade CERN Open Data portal version to 0.13.0 and change local
development server to use port 8000 in order to avoid conflicts with
port 5000 on macOS.

Additionally, plug the MQ service back to avoid unnecessary slow-downs
when accessing the detailed record pages. The MQ is needed by the
invenio-stats module that came in as part of the upgrade.